[cli] move Main to cli module
