+++ s3thorp-cli/src/main/scala/net/kemitix/s3thorp/cli/Main.scala
@@ -1,10 +1,11 @@
-package net.kemitix.s3thorp
+package net.kemitix.s3thorp.cli
 
 import java.nio.file.Paths
 
 import cats.effect.ExitCase.{Canceled, Completed, Error}
 import cats.effect.{ExitCode, IO, IOApp}
 import net.kemitix.s3thorp.awssdk.S3Client
+import net.kemitix.s3thorp.{Config, Logging, ParseArgs, Sync}
 
 object Main extends IOApp with Logging {
